requirements.txt
asgiref==3.11.0
Django==6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==12.0.0
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.5
tzdata==2025.3
